Move icell8_utils module -> icell8.utils.
